# openapi-spec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/projects/{id}": {
            "get": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "getProjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Project"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/saveOnBoardProject": {
            "post": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "createProject",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Project"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "getProjectByEmpId",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/activeProjectsForInvoiceByEmployee": {
            "get": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "activeProjectsForInvoiceByEmployee",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/activeProjectsForInvoice": {
            "get": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "activeProjectsForInvoiceByEmployee_1",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/activeProjectsForInvoice": {
            "get": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "activeProjectsForInvoiceByEmployee_2",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/activeProjectsForInvoiceByEmployee": {
            "get": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "activeProjectsForInvoiceByEmployee_3",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/activeProjects": {
            "get": {
                "tags": [
                    "project-controller"
                ],
                "operationId": "getAllActiveProjects_1",
                "parameters": [
                    {
                        "name": "endDate",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "selectedDate",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "type": "object",
                "properties": {
                    "addressId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "address": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zipCode": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Assignment": {
                "type": "object",
                "properties": {
                    "assignmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "projectId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "assignmentType": {
                        "type": "string"
                    },
                    "assignmentTaxType": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "wage": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string"
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string"
                    },
                    "assignmentNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notes"
                        }
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Customer": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerCompanyName": {
                        "type": "string"
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerPhone": {
                        "type": "string"
                    },
                    "customerEmail": {
                        "type": "string"
                    },
                    "ein": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    },
                    "customerStatus": {
                        "type": "string"
                    },
                    "customerStartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerEndDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerContactEmail": {
                        "type": "string"
                    },
                    "customerType": {
                        "type": "string"
                    },
                    "customerAddress": {
                        "type": "string"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Employee": {
                "type": "object",
                "properties": {
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "emailId": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "dob": {
                        "type": "string"
                    },
                    "ssn": {
                        "type": "string"
                    },
                    "taxTerm": {
                        "type": "string"
                    },
                    "referredBy": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "designation": {
                        "type": "string"
                    },
                    "employmentType": {
                        "type": "string"
                    },
                    "employeeType": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "primarySkills": {
                        "type": "string"
                    },
                    "secondarySkills": {
                        "type": "string"
                    },
                    "workCity": {
                        "type": "string"
                    },
                    "workCountry": {
                        "type": "string"
                    },
                    "resourceType": {
                        "type": "string"
                    },
                    "employeeDept": {
                        "type": "string"
                    },
                    "annualPay": {
                        "type": "number",
                        "format": "float"
                    },
                    "payrollStart": {
                        "type": "string",
                        "format": "date"
                    },
                    "everifyStatus": {
                        "type": "string"
                    },
                    "insurance": {
                        "type": "string"
                    },
                    "companyName": {
                        "type": "string"
                    },
                    "address": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Address"
                        }
                    },
                    "project": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "visas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Visa"
                        }
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    },
                    "i9": {
                        "type": "string"
                    },
                    "paf": {
                        "type": "string"
                    }
                }
            },
...
            "Notes": {
                "type": "object",
                "properties": {
                    "noteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "notesType": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Project": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/Customer"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    },
                    "client": {
                        "type": "string"
                    },
                    "billRates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Wage"
                        }
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "invoiceTerm": {
                        "type": "string"
                    },
                    "paymentTerm": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "assignments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Assignment"
                        }
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "vendorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Visa": {
                "type": "object",
                "properties": {
                    "visaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "visaCategory": {
                        "type": "string"
                    },
                    "receiptNumber": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "jobTitle": {
                        "type": "string"
                    },
                    "lcaNumber": {
                        "type": "string"
                    },
                    "socCode": {
                        "type": "string"
                    },
                    "client": {
                        "type": "string"
                    },
                    "vendor": {
                        "type": "string"
                    },
                    "jobLocation": {
                        "type": "string"
                    },
                    "jobLocation2": {
                        "type": "string"
                    },
                    "lcaWage": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string"
                    },
                    "filingType": {
                        "type": "string",
                        "enum": [
                            "TRANSFER",
                            "CONSULAR",
                            "COS"
                        ]
                    },
                    "visaSubCategory": {
                        "type": "string",
                        "enum": [
                            "TRANSFER",
                            "CAP",
                            "TRANSFER_INDIA"
                        ]
                    },
                    "filingYear": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    },
                    "lca": {
                        "$ref": "#/components/schemas/LCA"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Wage": {
                "type": "object",
                "properties": {
                    "wageId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "wage": {
                        "type": "number",
                        "format": "float"
                    },
                    "wageType": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "wageNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notes"
                        }
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
        }
    }
}